requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
openpyxl==3.1.5
html5lib==1.1
beautifulsoup4==4.12.3
chardet==5.2.0
zipfile36==0.1.3